added unit test - test_get_model_with_dataitem moved test_get_model_without_suffix to test_model.py
